start abstracting out of ztl
