+++ tests/test_io.py
@@ -1,10 +1,10 @@
 # test s3 and sql data IO
-# TODO abstract these functions out of ztl directory
 from src.postgres_utils import get_table_columns
 from src.digital_ocean_utils import get_datatset_config
-from src.ztl.helpers import DATASET_QAQC_DB_SCHEMA
 
-TEST_DATASET_NAME = DATASET_QAQC_DB_SCHEMA
+# DEV temporary
+from src.source_report_utils import DATASET_QAQC_DB_SCHEMA
+
 TEST_DATA_SOURCE_NAME = "dcp_zoningmapamendments"
 TEST_DATA_SOURCE_VERSION = "20230404"
 TEST_DATA_SOURCE_COLUMNS = [
@@ -31,9 +31,7 @@ def test_dataset_config():
 
 def test_get_table_columns():
     columns = get_table_columns(
-        # table_schema="db_zoningtaxlots",
-        # table_name="dcp_zoningmapamendments_20230306",
-        table_schema=TEST_DATASET_NAME,
+        table_schema=DATASET_QAQC_DB_SCHEMA,
         table_name=f"{TEST_DATA_SOURCE_NAME}_{TEST_DATA_SOURCE_VERSION}",
     )
     assert columns == TEST_DATA_SOURCE_COLUMNS
